==> 04_inside_survival.Rmd <==
---
title: "A peek inside the black box of apprenticeship"
output:
  flexdashboard::flex_dashboard:
    orientation: rows
    vertical_layout: fill
runtime: shiny
resource_files:
- out/nested_with_summary.rds
- out/time_and_status.rds
- out/state_prob_plots.rds
---

```{css}
table {
    font-size: 15px; /* Adjust the size as needed */
}
```

```{r setup, include=FALSE}
#libraries----------------------------------
library(flexdashboard)
library(tidyverse)
library(here)
library(survival)
library(shinyalert)
library(conflicted)
conflicts_prefer(dplyr::filter)
#functions--------------------------------------
source("functions.R")
#read in the data---------------------------------------
state_prob_plots <- read_rds(here("out","state_prob_plots.rds"))
nested_with_summary <- read_rds(here("out", "nested_with_summary.rds"))|>
  mutate(trade_desc=str_remove_all(trade_desc,",")) #commas in facet labels screw up annotation
t_and_s <- read_rds(here("out", "time_and_status.rds"))|>
  mutate(trade_desc=str_remove_all(trade_desc,","))#commas in facet labels screw up annotation
levs <- unique(t_and_s$level)
trades <- unique(t_and_s$trade_desc)
```

# By Trade

## Inputs {.sidebar}

```{r}
selectInput(
  "trade",
  "Select a trade: ",
  trades,
  selected = "Carpenter"
)


base_table <- reactive({
  nested_with_summary|>
    filter(trade_desc==input$trade,
           level!="journeyperson")|>
    ungroup()|>
    select(level, era, probability, years)|>
    arrange(era)
})

totals <- reactive({
  base_table()|>
    group_by(era)|>
    summarize(level="Journey-person",
              probability=prod(probability), 
              years=sum(years))
})

trade_summaries <- reactive({
  bind_rows(base_table(), totals())
})
#error message when state probability plot not present. 
trade_state_plots <- reactive({
  validate(
      need(input$trade %in% unique(state_prob_plots$trade_desc), "For this trade there are not enough data points to create a meaningful state probability plot.")
  )
  state_prob_plots|>
    filter(trade_desc==input$trade)
})

```

-  There are 5 tabs on this page:
-   The first tab gives the completion curves: the probability of completing a level in terms of time since registration (for level 1 and journeyperson) or completion of the previous level (all other levels).
-   The second tab is a table of the completion probabilities.
-   The third tab is a table of the expected durations.
-   The fourth tab describes how to go from survival curves to state probabilities.
-   The fifth tab is an area plot of the state probabilities.
-   The data is evenly split into 3 eras based on the date of registration. 

Row {.tabset}
-------------------------------------
### Completion Curves

```{r, fig.retina=2}
renderPlot({
  trade_surv(input$trade, t_and_s)
})
```

### Probability of Level Completion

```{r}
DT::renderDT(
  trade_summaries()|>
    select(-years)|>
    pivot_wider(names_from = era, values_from = probability)|>
    my_dt()
)
```

### Expected Level Duration (years)

```{r}
DT::renderDT(
  trade_summaries()|>
    select(-probability)|>
    pivot_wider(names_from = era, values_from = years)|>
    my_dt()
)
```

### Survival -> State Probabilities

-   In order to estimate state probabilities we need to estimate the survival curves for each level from the time of registration: $survival(level)$  
-   This is different than the completion curves, where time was measured from completion of preceding level (with the exception of level 1 and journey-person).
-   Note the following:
$$
\begin{eqnarray}
1-survival(completion)&=&probability(completion)\\
1-survival(level4) &=& probability(completion~or~level4)\\
1-survival(level3) &=& probability(completion~or~level4~or~level3)\\
1-survival(level2) &=& probability(completion~or~level4~or~level3~or~level2)\\
1-survival(level1) &=& probability(completion~or~level4~or~level3~or~level2~or~level1)\\
survival(level1) &=& probability(no~levels)
\end{eqnarray}
$$
-   We can utilize the recursive nature of the above probabilities to derive the state probabilities:

$$
\begin{eqnarray}
probability(completion)&=&1-survival(completion)\\
probability(level4) &=& survival(completion)-survival(level4)\\
probability(level3) &=& survival(level4)-survival(level3)\\
probability(level2) &=& survival(level3)-survival(level2)\\
probability(level1) &=& survival(level2)-survival(level1)\\
probability(no~levels) &=& survival(level1) 
\end{eqnarray}
$$

-   Note that if the survival curves feature large jumps then the *estimated* state probabilities can be negative (obviously the *actual* state probabilities must be weakly positive.)
-   Large jumps in the survival curves are more likely when sample sizes are small.  
-   State probability plots are only provided for the trades where all state probabilities are weakly positive.


### State probability plots

```{r, fig.retina=2}
renderPlot({
  trade_state_plots()[["state_prob_plot"]]
})
```

# By Level

## Inputs {.sidebar}

```{r}
selectInput(
  "level",
  "Select a level: ",
  levs,
  selected = levs[1]
)

level_summaries <- reactive({
  nested_with_summary|>
    filter(level==input$level)|>
    ungroup()|>
    select(trade_desc, era, probability, years)|>
    arrange(era)
})
```

-  There are 3 tabs on this page for `r renderUI({input$level})`.
-   The first tab shows the probability of completing a level in terms of time since registration (for level 1 and journeyperson) or completion of the previous level (all other levels).
-   The second tab is a table of the completion probabilities.
-   The third tab is a table of the expected durations.

Note:

- Prior to the program change in April 2014, the Transport Trailer Technician program did not require any technical training, resulting in a value of 0 for technical training level.
- Prior to August 2012, Residential Steep Roofer's technical training was Roofer's 3 levels of TT, then it changed to 1 level of Residential Steep Roofer.

Row {.tabset}
-------------------------------------
### Completion Curves


```{r, fig.retina=2}
renderPlot({
  level_surv(input$level, t_and_s)
})
```

### Probability of Level Completion

```{r}
DT::renderDT(
  level_summaries()|>
    select(-years)|>
    pivot_wider(names_from = era, values_from = probability)|>
    my_dt()
)
```

### Expected Level Duration (years)

```{r}
DT::renderDT(
  level_summaries()|>
    select(-probability)|>
    pivot_wider(names_from = era, values_from = years)|>
    my_dt()
)
```
